Fix migrations to use bigint data type instead of integer for reference to adopter_foster_account id and change references in code to adopter account to use the new model/method name

+++ db/migrate/20240318174108_add_adopter_foster_account_id_to_adopter_foster_profiles.rb
@@ -1,6 +1,6 @@
 class AddAdopterFosterAccountIdToAdopterFosterProfiles < ActiveRecord::Migration[7.0]
   def change
-    add_column :adopter_foster_profiles, :adopter_foster_account_id, :integer
+    add_column :adopter_foster_profiles, :adopter_foster_account_id, :bigint
     add_index :adopter_foster_profiles, :adopter_foster_account_id
   end
 end
